+++ apps/docs/src/examples/core/three-arcade-game/ArcadeMachineScene.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { InteractiveObject, Three2, useFrame, useTexture, useThrelte } from '@threlte/core'
+	import { InteractiveObject, Three, useFrame, useTexture, useThrelte } from '@threlte/core'
 	import { Environment, useCursor, useGltf } from '@threlte/extras'
 	import { onMount } from 'svelte'
 	import { cubicInOut } from 'svelte/easing'
@@ -252,10 +252,10 @@
 
 <svelte:window on:keydown={onKeyDown} on:keyup={onKeyUp} />
 
-<Three2 type={Scene} bind:ref={$arcadeMachineScene} background={new Color(0x020203)}>
+<Three type={Scene} bind:ref={$arcadeMachineScene} background={new Color(0x020203)}>
 	<Environment path="/hdr/" files="shanghai_riverside_1k.hdr" />
 
-	<Three2
+	<Three
 		position.x={$cameraTargetPos.x}
 		position.y={$cameraTargetPos.y}
 		position.z={$cameraTargetPos.z}
@@ -265,7 +265,7 @@
 	/>
 
 	{#if $orbitControls}
-		<Three2
+		<Three
 			type={PerspectiveCamera}
 			position.x={20}
 			position.y={20}
@@ -274,10 +274,10 @@
 			makeDefault
 			let:ref={camera}
 		>
-			<Three2 type={OrbitControls} args={[camera, renderer?.domElement]} />
-		</Three2>
+			<Three type={OrbitControls} args={[camera, renderer?.domElement]} />
+		</Three>
 	{:else}
-		<Three2
+		<Three
 			type={PerspectiveCamera}
 			position.x={$cameraPos.x}
 			position.y={$cameraPos.y}
@@ -291,27 +291,27 @@
 
 	{#if nodes && materials}
 		<!-- Generated by gltfjsx -->
-		<Three2 type={Group} rotation.y={DEG2RAD * 180}>
-			<Three2
+		<Three type={Group} rotation.y={DEG2RAD * 180}>
+			<Three
 				type={Mesh}
 				geometry={nodes.BodyMesh.geometry}
 				material={materials['machine body main']}
 				position={[0.2755, 0, 0]}
 			/>
-			<Three2
+			<Three
 				type={Mesh}
 				geometry={nodes.LeftCover.geometry}
 				material={materials['machine body outer']}
 				position={[0.3, 1.2099, -0.1307]}
 			/>
-			<Three2
+			<Three
 				type={Mesh}
 				geometry={nodes.RightCover.geometry}
 				material={materials['machine body outer']}
 				position={[-0.3, 1.2099, -0.1307]}
 				scale={[-1, 1, 1]}
 			/>
-			<Three2
+			<Three
 				type={Mesh}
 				geometry={nodes.ScreenFrame.geometry}
 				material={materials['screen frame']}
@@ -320,37 +320,37 @@
 
 			<!-- Joystick -->
 
-			<Three2
+			<Three
 				type={Mesh}
 				geometry={nodes.joystick_base.geometry}
 				material={materials['joystick base']}
 				position={[0.1336, 0.9611, -0.1976]}
 				rotation={[-0.1939, 0, 0]}
 			/>
-			<Three2
+			<Three
 				type={Mesh}
 				geometry={nodes.joystick_stick_application.geometry}
 				material={materials['joystick base']}
 				position={[0.1336, 0.9653, -0.1984]}
 				rotation={[-0.1939, 0, $rotationStick]}
 			>
-				<Three2
+				<Three
 					type={Mesh}
 					geometry={nodes.joystick_stick.geometry}
 					material={materials['joystick stick']}
 					position={[0, -0.0145, 0.0001]}
 				>
-					<Three2
+					<Three
 						type={Mesh}
 						geometry={nodes.joystick_cap.geometry}
 						material={materials['joystick cap']}
 						position={[-0.0001, 0.1126, -0.0005]}
 						material.envMapIntensity={0.5}
 					/>
-				</Three2>
-			</Three2>
+				</Three>
+			</Three>
 
-			<Three2
+			<Three
 				type={Mesh}
 				geometry={nodes.Main_Button_Enclosure.geometry}
 				material={materials['joystick base']}
@@ -358,7 +358,7 @@
 				rotation={[-0.1801, 0, 0]}
 				scale={0.9409}
 			>
-				<Three2
+				<Three
 					type={Mesh}
 					geometry={nodes.Main_Button.geometry}
 					material={materials['joystick cap']}
@@ -366,10 +366,10 @@
 					rotation={[0.192, 0, 0]}
 					scale={0.724}
 				/>
-			</Three2>
+			</Three>
 
 			<!-- The screen itself gets a special treatment -->
-			<Three2
+			<Three
 				type={Mesh}
 				geometry={nodes.Screen.geometry}
 				position={[0, 1.3774, 0.1447]}
@@ -383,7 +383,7 @@
 					interactive
 					on:click={onScreenClick}
 				/>
-				<Three2
+				<Three
 					type={MeshStandardMaterial}
 					metalness={0.9}
 					roughness={0.2}
@@ -395,13 +395,13 @@
 					emissiveIntensity={1.2}
 					envMapIntensity={0}
 				/>
-			</Three2>
-		</Three2>
+			</Three>
+		</Three>
 	{/if}
 
 	<!-- This PointLight replicates the light emitted by the screen -->
 	{#if nodes}
-		<Three2
+		<Three
 			type={PointLight}
 			args={['black']}
 			position.y={1.376583185239323}
@@ -414,11 +414,11 @@
 		/>
 	{/if}
 
-	<Three2 type={AmbientLight} intensity={8} color={$averageScreenColor} />
-	<Three2 type={AmbientLight} intensity={$whiteAmbientLightIntensity} color={'white'} />
+	<Three type={AmbientLight} intensity={8} color={$averageScreenColor} />
+	<Three type={AmbientLight} intensity={$whiteAmbientLightIntensity} color={'white'} />
 
 	<!-- Red light -->
-	<Three2
+	<Three
 		type={DirectionalLight}
 		intensity={$redLightIntensity}
 		color="#F67F55"
@@ -428,7 +428,7 @@
 	/>
 
 	<!-- Blue light -->
-	<Three2
+	<Three
 		type={DirectionalLight}
 		intensity={$blueLightIntensity}
 		position.x={2.2}
@@ -438,7 +438,7 @@
 	/>
 
 	<!-- White light -->
-	<Three2
+	<Three
 		type={DirectionalLight}
 		intensity={$whiteLightIntensity}
 		position.x={-1}
@@ -448,8 +448,8 @@
 	/>
 
 	<!-- Floor -->
-	<Three2 type={Mesh}>
-		<Three2 type={CylinderGeometry} args={[1, 1, 0.04, 64]} />
-		<Three2 type={MeshStandardMaterial} color={'#0f0f0f'} />
-	</Three2>
-</Three2>
+	<Three type={Mesh}>
+		<Three type={CylinderGeometry} args={[1, 1, 0.04, 64]} />
+		<Three type={MeshStandardMaterial} color={'#0f0f0f'} />
+	</Three>
+</Three>
