+++ apps/docs/src/examples/core/three-arcade-game/game/Arena.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { Three2 } from '@threlte/core'
+	import { Three } from '@threlte/core'
 	import { Collider } from '@threlte/rapier'
 	import { LineBasicMaterial, Mesh, MeshBasicMaterial, PlaneGeometry } from 'three'
 	import { DEG2RAD } from 'three/src/math/MathUtils'
@@ -17,59 +17,59 @@
 </script>
 
 <!-- BACKGROUND GRID -->
-<Three2
+<Three
 	type={CustomGridHelper}
 	args={[arenaWidth, arenaWidth, arenaHeight, arenaWidth]}
 	position.y={-0.5}
 >
-	<Three2 type={LineBasicMaterial} color="green" transparent opacity={0.1} />
-</Three2>
+	<Three type={LineBasicMaterial} color="green" transparent opacity={0.1} />
+</Three>
 
 <!-- LEFT GRID -->
-<Three2
+<Three
 	type={CustomGridHelper}
 	args={[arenaDepth, arenaDepth, arenaHeight, arenaHeight]}
 	rotation.z={90 * DEG2RAD}
 	position.x={(arenaWidth / 2) * -1}
 >
-	<Three2 type={LineBasicMaterial} color="green" transparent opacity={sideGridOpacity} />
+	<Three type={LineBasicMaterial} color="green" transparent opacity={sideGridOpacity} />
 
-	<Three2 type={Mesh} rotation.x={90 * DEG2RAD}>
-		<Three2 type={PlaneGeometry} args={[arenaDepth, arenaHeight]} />
-		<Three2 type={MeshBasicMaterial} color="green" transparent opacity={$leftOpacity} />
-	</Three2>
-</Three2>
+	<Three type={Mesh} rotation.x={90 * DEG2RAD}>
+		<Three type={PlaneGeometry} args={[arenaDepth, arenaHeight]} />
+		<Three type={MeshBasicMaterial} color="green" transparent opacity={$leftOpacity} />
+	</Three>
+</Three>
 
 <!-- RIGHT GRID -->
-<Three2
+<Three
 	type={CustomGridHelper}
 	args={[arenaDepth, arenaDepth, arenaHeight, arenaHeight]}
 	rotation.z={90 * DEG2RAD}
 	position.x={arenaWidth / 2}
 >
-	<Three2 type={LineBasicMaterial} color="green" transparent opacity={sideGridOpacity} />
+	<Three type={LineBasicMaterial} color="green" transparent opacity={sideGridOpacity} />
 
-	<Three2 type={Mesh} rotation.x={-90 * DEG2RAD}>
-		<Three2 type={PlaneGeometry} args={[arenaDepth, arenaHeight]} />
-		<Three2 type={MeshBasicMaterial} color="green" transparent opacity={$rightOpacity} />
-	</Three2>
-</Three2>
+	<Three type={Mesh} rotation.x={-90 * DEG2RAD}>
+		<Three type={PlaneGeometry} args={[arenaDepth, arenaHeight]} />
+		<Three type={MeshBasicMaterial} color="green" transparent opacity={$rightOpacity} />
+	</Three>
+</Three>
 
 <!-- TOP GRID -->
-<Three2
+<Three
 	type={CustomGridHelper}
 	args={[arenaDepth, arenaDepth, arenaHeight, arenaHeight]}
 	rotation.y={90 * DEG2RAD}
 	rotation.x={90 * DEG2RAD}
 	position.z={(arenaHeight / 2) * -1}
 >
-	<Three2 type={LineBasicMaterial} color="green" transparent opacity={sideGridOpacity} />
+	<Three type={LineBasicMaterial} color="green" transparent opacity={sideGridOpacity} />
 
-	<Three2 type={Mesh} rotation.x={-90 * DEG2RAD}>
-		<Three2 type={PlaneGeometry} args={[arenaDepth, arenaHeight]} />
-		<Three2 type={MeshBasicMaterial} color="green" transparent opacity={$topOpacity} />
-	</Three2>
-</Three2>
+	<Three type={Mesh} rotation.x={-90 * DEG2RAD}>
+		<Three type={PlaneGeometry} args={[arenaDepth, arenaHeight]} />
+		<Three type={MeshBasicMaterial} color="green" transparent opacity={$topOpacity} />
+	</Three>
+</Three>
 
 <!-- LEFT COLLIDER -->
 <Collider
