+++ packages/core/src/lib/index.ts
@@ -61,8 +61,6 @@ export { default as DisposableObject } from './internal/DisposableObject.svelte'
 
 // <Three> component
 export { default as Three } from './three/Three.svelte'
-export { default as Three2 } from './three/Three2.svelte'
-export { extend } from './three/extend'
 
 // hooks
 export { useFrame } from './hooks/useFrame'
