+++ packages/core/src/lib/three/Three.svelte
@@ -1,98 +1,152 @@
 <script lang="ts">
   import { getContext, setContext } from 'svelte'
-  import type { Writable } from 'svelte/store'
-  import { Object3D } from 'three'
-  import { useThrelteRoot } from '../hooks/useThrelteRoot'
+  import { writable, type Writable } from 'svelte/store'
+  import type { Object3D } from 'three'
+  import type { ConditionalKeys, Primitive } from 'type-fest'
+  import { useThrelte } from '../hooks/useThrelte'
   import DisposableObject from '../internal/DisposableObject.svelte'
   import SceneGraphObject from '../internal/SceneGraphObject.svelte'
-  import { createObjectStore } from '../lib/createObjectStore'
   import type { DisposableThreeObject } from '../types/components'
-  import { useExtensions } from './extend'
-  import { prepareInstance } from './utils/prepare'
-  import { applyProps } from './utils/props'
-  import { useAttach } from './utils/useAttach'
-
-  type ThrelteThreeParentContext = Writable<any | undefined>
-
-  const parent = getContext<ThrelteThreeParentContext>('threlte-hierarchical-parent-context')
-
-  const extended = useExtensions()
-
-  type T = $$Generic<keyof typeof extended>
-
-  export let type: T
+  import type { Attach } from './lib/types'
+  import { useAttach } from './lib/useAttach'
+  import { useCamera } from './lib/useCamera'
+  import { useProps } from './lib/useProps'
 
   type AnyClass = new (...args: any) => any
-  type ThreeClass = typeof extended[typeof type] extends new (...args: any) => any
-    ? typeof extended[typeof type]
-    : never
-  type ThreeClassInstance = ThreeClass extends never ? never : InstanceType<ThreeClass>
-  type Args = ThreeClass extends AnyClass ? ConstructorParameters<ThreeClass> : never
+  type AnyFn = (...args: any) => any
 
-  export let args: Args | any[] | undefined = undefined
-  export let attach:
-    | string
-    | ((parent: any, self: ThreeClassInstance) => (() => void) | void)
-    | undefined = undefined
+  /**
+   * We hold a list of prop keys that should be ommited from the object props
+   * that are infered by the provided type.
+   */
+  type OmittedPropKeys =
+    | 'type'
+    | 'args'
+    | 'attach'
+    | 'manual'
+    | 'makeDefault'
+    | 'id'
+    | 'children'
+    | `is${string}` // isMesh, isObject3D, etc
+    | 'parent'
+    | 'uuid'
+    | 'name'
 
-  type $$Props = {
-    js: T
-    args?: Args | any[]
-    attach?: string | ((parent: any, self: any) => (() => void) | void)
-    object?: ThreeClassInstance
-  } & Record<
-    string,
-    any
-  > & {
-			/* prettier-ignore */
-			// let's add types for object setters and properties
-      [K in keyof ThreeClassInstance]?:
-			// if a property has a function "set" …
-			ThreeClassInstance[K] extends { set: (...args: any[]) => any }
-				// … then we can use it as a type, setScalar is
-				// included as it's always the first argument to
-				// the properties set function
-        ? Parameters<ThreeClassInstance[K]['set']> | Parameters<ThreeClassInstance[K]['set']>[0]
-				// if the property does not have a "set" function,
-				// and is a function …
-        : ThreeClassInstance[K] extends (...args: any[]) => any
-				// we don't want it to be a property
-        ? never
-				// otherwise, it's a property with just a value
-        : ThreeClassInstance[K]
-    }
+  // Basic Prop Types
+  type Type = $$Generic
+  type Instance = Type extends AnyClass ? InstanceType<Type> : Type
+  type Manual = Instance extends { isCamera: true } ? boolean : never
+  type MakeDefault = Instance extends { isCamera: true } ? boolean : never
+  type Args = Type extends AnyClass ? ConstructorParameters<Type> : never
 
-  const makeBaseInstance = (args: any): ThreeClassInstance => {
-    return (
-      args ? new (extended[type] as any)(...args) : new (extended[type] as any)()
-    ) as ThreeClassInstance
+  /**
+   * Record<string, any> allows the use of any prop as there could be pierced
+   * props ("position.x") which are not picked up by the following types.
+   */
+  type AnyProps = Record<string, any>
+  type ClassProps = Type extends AnyClass
+    ? {
+        /**
+         * Constructor Args passed to the constructor if a class is passed to "type"
+         */
+        args?: Args | any[]
+      }
+    : {}
+  type BaseProps = {
+    type: Type
+    attach?: Attach<Type>
+  }
+  type CameraProps = Instance extends { isCamera: true }
+    ? {
+        /**
+         * By default, threlte will update the cameras aspect ratio or frustum
+         * when the canvas is resized. If you want to manually control the
+         * camera, set this to true.
+         * @default true
+         */
+        manual?: Manual
+        /**
+         * Sets the camera as the default camera for the scene.
+         * @default false
+         */
+        makeDefault?: MakeDefault
+      }
+    : {}
+  type InstanceProps = Omit<
+    Instance extends Primitive
+      ? {}
+      : {
+          [K in keyof Instance]?: Instance[K] extends { set: (...args: any[]) => any }
+            ? Parameters<Instance[K]['set']> | Parameters<Instance[K]['set']>[0]
+            : Instance[K] extends AnyFn
+            ? never
+            : Instance[K]
+        },
+    ConditionalKeys<Instance, AnyFn> | OmittedPropKeys
+  >
+
+  type AllProps = AnyProps & BaseProps & CameraProps & InstanceProps & ClassProps
+  type $$Props = AllProps
+
+  export let type: Type
+  export let args: Args = undefined as Args
+  export let attach: Attach<Type> = undefined as Attach<Type>
+  export let manual: Manual = undefined as unknown as Manual
+  export let makeDefault: MakeDefault = undefined as unknown as MakeDefault
+
+  type ThrelteThreeParentContext = Writable<any | undefined>
+  const parent = getContext<ThrelteThreeParentContext>('threlte-hierarchical-parent-context')
+
+  // Type Gaurds
+  const isClass = (type: any): type is AnyClass => {
+    return typeof type === 'function' && /^\s*class\s+/.test(type.toString())
+  }
+  const argsIsConstructorParameters = (args: any): args is ConstructorParameters<AnyClass> => {
+    return Array.isArray(args)
   }
 
-  if (typeof extended[type] !== 'function') {
-    throw new Error(`extended[${type}] is not a constructor`)
+  // We can't create the object in a reactive statement due to providing context
+  export let ref = (
+    isClass(type) && argsIsConstructorParameters(args)
+      ? new type(...(args as any)) // TODO: fix this any
+      : isClass(type)
+      ? new type()
+      : type
+  ) as Instance
+  let initialized = false
+  $: if (initialized) {
+    ref = (
+      isClass(type) && argsIsConstructorParameters(args)
+        ? new type(...(args as any)) // TODO: fix this any
+        : isClass(type)
+        ? new type()
+        : type
+    ) as Instance
+  } else {
+    initialized = true
   }
-  const instance = prepareInstance(makeBaseInstance(args), type, $parent)
-  $: instance._threlte_three.parent = $parent
-
-  export const object = instance as ThreeClassInstance
-
-  $: applyProps(instance, $$props)
-
-  const objectStore = createObjectStore(instance as any)
-  $: $objectStore = instance
-
+  const objectStore = writable(ref)
+  $: objectStore.set(ref)
   setContext<ThrelteThreeParentContext>('threlte-hierarchical-parent-context', objectStore)
 
-  const { setCamera } = useThrelteRoot()
-  const isCamera = type.endsWith('Camera')
-  $: if (isCamera) {
-    setCamera(instance as any)
-  }
+  // Props
+  $: props = useProps()
+  $: props.updateProps(ref, $$restProps, {
+    manualCamera: manual
+  })
 
-  useAttach(instance, attach)
+  // Camera
+  const { size } = useThrelte()
+  $: camera = useCamera()
+  $: camera.update(ref, $size, manual)
+  $: camera.makeDefaultCamera(ref, makeDefault)
+
+  // Attachment
+  $: attachment = useAttach(ref, attach)
+  $: attachment.update(ref, $parent, attach)
 
   const extendsObject3D = (object: any): object is Object3D => {
-    return (object as any) instanceof Object3D
+    return !!(object as any).isObject3D
   }
 
   const isDisposableObject = (object: any): object is DisposableThreeObject => {
@@ -100,14 +154,14 @@
   }
 </script>
 
-{#if isDisposableObject(object)}
-  <DisposableObject {object} />
+{#if isDisposableObject(ref)}
+  <DisposableObject object={ref} />
 {/if}
 
-{#if extendsObject3D(object)}
-  <SceneGraphObject {object}>
-    <slot {object} />
+{#if extendsObject3D(ref)}
+  <SceneGraphObject object={ref}>
+    <slot {ref} />
   </SceneGraphObject>
 {:else}
-  <slot {object} />
+  <slot {ref} />
 {/if}
