+++ packages/core/src/lib/three/lib/resolve.ts
