+++ packages/core/src/lib/three/lib/types.ts
