+++ packages/core/src/lib/three/lib/useProps.ts
